Two tanks for one user

+++ src/main/java/tank/dto/TankDto.java
@@ -5,66 +5,66 @@ import java.io.Serializable;
 public class TankDto implements Serializable {
     private static final long serialVersionUID = 8038539938717817115L;
     int id;
-    private int speed = 5;
-    public int alpha = 0;
-    int deltaX = 0;
-    int deltaY = 0;
-    int deltaAlpha = 0;
-    int speedAlpha = 2;
+    private float speed = 5;
+    public float alpha = 0;
+    float deltaX = 0;
+    float deltaY = 0;
+    float deltaAlpha = 0;
+    float speedAlpha = 2;
     private ToreDto tore;
 
     public TankDto(int id) {
         this.id = id;
     }
 
-    private int X = 0;
-    private int Y = 0;
+    private float X = 0;
+    private float Y = 0;
 
-    public int getSpeed() {
+    public float getSpeed() {
         return speed;
     }
 
-    public void setSpeed(int speed) {
+    public void setSpeed(float speed) {
         this.speed = speed;
     }
 
-    public int getAlpha() {
+    public float getAlpha() {
         return alpha;
     }
 
-    public void setAlpha(int alpha) {
+    public void setAlpha(float alpha) {
         this.alpha = alpha;
     }
 
-    public int getDeltaX() {
+    public float getDeltaX() {
         return deltaX;
     }
 
-    public void setDeltaX(int deltaX) {
+    public void setDeltaX(float deltaX) {
         this.deltaX = deltaX;
     }
 
-    public int getDeltaY() {
+    public float getDeltaY() {
         return deltaY;
     }
 
-    public void setDeltaY(int deltaY) {
+    public void setDeltaY(float deltaY) {
         this.deltaY = deltaY;
     }
 
-    public int getDeltaAlpha() {
+    public float getDeltaAlpha() {
         return deltaAlpha;
     }
 
-    public void setDeltaAlpha(int deltaAlpha) {
+    public void setDeltaAlpha(float deltaAlpha) {
         this.deltaAlpha = deltaAlpha;
     }
 
-    public int getSpeedAlpha() {
+    public float getSpeedAlpha() {
         return speedAlpha;
     }
 
-    public void setSpeedAlpha(int speedAlpha) {
+    public void setSpeedAlpha(float speedAlpha) {
         this.speedAlpha = speedAlpha;
     }
 
@@ -76,19 +76,19 @@ public class TankDto implements Serializable {
         this.id = id;
     }
 
-    public int getX() {
+    public float getX() {
         return X;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         X = x;
     }
 
-    public int getY() {
+    public float getY() {
         return Y;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         Y = y;
     }
